requirements.txt
python-dotenv
openai
python-telegram-bot
Flask